Rename CacheTypeIdentifierUtil
